Parentheses and better looking tests

+++ src/reprs.rs
@@ -1,6 +1,6 @@
 use crate::error::{Error, Result};
 
-pub fn parse_num(num_str: &str) -> Result<i64> {
+pub fn parse_num(num_str: &str) -> Result<i128> {
     let mut radix: Option<u32> = None;
     let mut slice = num_str;
     if slice.starts_with("0x") {
@@ -17,10 +17,10 @@ pub fn parse_num(num_str: &str) -> Result<i64> {
         radix = Some(2);
     }
     if let Some(radix) = radix {
-        i64::from_str_radix(slice, radix).map_err(|_| Error::NumParseError(num_str.to_string()))
+        i128::from_str_radix(slice, radix).map_err(|_| Error::NumParseError(num_str.to_string()))
     } else {
         [10, 16].iter()
-            .find_map(|&radix| i64::from_str_radix(slice, radix).ok())
+            .find_map(|&radix| i128::from_str_radix(slice, radix).ok())
             .ok_or(Error::NumParseError(num_str.to_string()))
     }
 }
